fix(estimation): replication is opt-in; JKn dead ends warn at construction
method=None resolved to replication whenever the design carried replicate
weights and no stratum/psu, and to Taylor otherwise. That made the estimator
depend on inputs the estimator never reads -- replication consumes the
replicate columns and coefficients(), nothing else -- so declaring a single
design column silently moved the standard error across a 25% spread on
identical weights. It was worst for JKn, where stratum and psu are exactly
what svy needs to derive (n_h-1)/n_h: the one action that makes JKn usable
was the action that switched JKn off.

None is now the unstated Taylor default the signature already implied.
Falling through to Taylor alone would have reinstated the worse half of the
original bug, which is why the auto-detection existed -- without stratum or
psu linearization is SRS-like, df=199 against a true 7 on a 200-row file,
silently. That case now warns (TAYLOR_WITHOUT_DESIGN) and proceeds, for the
explicit method="taylor" spelling too.

Also:
- A declared JKn that cannot be derived -- no psu to count from, or
  unbalanced strata -- warns at Sample construction instead of building in
  silence and failing at a call site far from the cause. The MethodError
  stays lazy: such a Sample is still usable for Taylor and wrangling.
- That error names both remedies; it named only scale=, sending anyone whose
  file does carry psu off to hand-compute what svy would have derived.
- create_bs_wgts(kind="poisson") raised NameError rather than its intended
  message against an older svy-rs; the fallback never bound the symbol.
- MethodError.not_applicable no longer doubles the sentence-ending period.
- RepWeights.df is annotated float, matching what it has always stored.
  Widened rather than coerced: df feeds a t-quantile, which is defined for
  fractional df, and Satterthwaite-style effective df is fractional.

+++ packages/svy/src/svy/core/design.py
@@ -358,7 +358,7 @@ class Design:
         prefix: str | _MissingType = _MISSING,
         n_reps: int | _MissingType = _MISSING,
         fay_coef: float | _MissingType = _MISSING,
-        df: int | None | _MissingType = _MISSING,
+        df: float | None | _MissingType = _MISSING,
         padding: int | None | _MissingType = _MISSING,
         scale: float | Sequence[float] | None | _MissingType = _MISSING,
         rep_coefs: tuple[float, ...] | None | _MissingType = _MISSING,
